Rename FlamePy agent API to service session
