+++ docs/designs/RFE309-agent-client/FS.md
@@ -1,55 +1,55 @@
-# RFE309: Agent client
+# RFE309: Service session client
 
 ## Motivation
 
-Currently, users build AI Agents by directly re-using the core APIs. However, this approach can be confusing and inconvenient for end users. In this feature, a dedicated Agent client will be introduced to streamline and simplify the process of building Agent clients.
+Currently, users build AI agents by directly re-using the core APIs. However, this approach can be confusing and inconvenient for end users. In this feature, a dedicated service session client will be introduced to streamline and simplify the process of interacting with service sessions.
 
 ## Function Specification
 
-A new class, called `Agent`, will be added to the `flamepy.agent` module to provide a streamlined interface for building and interacting with agents.
+A new class, called `Session`, will be added to the `flamepy.service` module to provide a streamlined interface for interacting with service sessions.
 
 ### Initialization
 
-There are two ways to manage the lifecycle of an Agent: using the constructor directly or as a context manager.
+There are two ways to manage the lifecycle of a service session: using the constructor directly or as a context manager.
 
 - **Direct Instantiation:**  
-  When you create an agent using the constructor, you are responsible for explicitly closing it after use. This ensures that all resources are properly released.
+  When you create a service session using the constructor, you are responsible for explicitly closing it after use. This ensures that all resources are properly released.
 
   ```python
-  agent = Agent("test", ctx)
-  # ... use the agent ...
-  agent.close()
+  session = Session("test", ctx)
+  # ... use the session ...
+  session.close()
   ```
 
 - **Context Manager:**  
-  By using the agent as a context manager, resource management is handled automatically. The agent is closed when execution leaves the context block.
+  By using the session as a context manager, resource management is handled automatically. The session is closed when execution leaves the context block.
 
   ```python
-  with Agent("test", ctx) as agent:
-      # ... use the agent ...
+  with Session("test", ctx) as session:
+      # ... use the session ...
   ```
 
-### Invoking the Agent
+### Invoking the Service Session
 
-Use the `invoke()` method to interact with the agent. The request and response objects should exactly match the agent's entrypoint signature defined on the service side.
+Use the `invoke()` method to interact with the service session. The request and response objects should exactly match the service entrypoint signature defined on the service side.
 
 ```python
-resp = agent.invoke(req)
+resp = session.invoke(req)
 ```
 
-### Accessing the Agent Context
+### Accessing the Service Session Context
 
-The current agent context—such as a running conversation or environment state—can be retrieved via the `context()` method.
+The current service session context, such as a running conversation or environment state, can be retrieved via the `context()` method.
 
 ```python
-ctx = agent.context()
+ctx = session.context()
 ```
 
 ## Implementation detail
 
-1. The `Agent` class is essentially a thin wrapper around `core.Session`, where the Agent's name corresponds to the application's name.
-2. The agent's `context()` method returns the session's shared (common) data, accessed directly from `core.Session`.
-3. The agent's `invoke()` method delegates its call to `core.Session.invoke`, passing through the invocation to the underlying session mechanism.
+1. The `Session` class is essentially a thin wrapper around `core.Session`, where the service session name corresponds to the application's name.
+2. The service session's `context()` method returns the session's shared (common) data, accessed directly from `core.Session`.
+3. The service session's `invoke()` method delegates its call to `core.Session.invoke`, passing through the invocation to the underlying session mechanism.
 
 ## Use Cases
