+++ e2e/tests/test_agent.py
@@ -14,7 +14,7 @@ limitations under the License.
 import flamepy
 import pytest
 from flamepy import SessionState
-from flamepy.agent import Agent
+from flamepy.service import Session
 
 from e2e.api import TestContext, TestRequest
 from tests.utils import random_string
@@ -49,158 +49,158 @@ def setup_test_env():
 
 
 # ============================================================================
-# Agent API Tests
+# Service Session API Tests
 # ============================================================================
 
 
-def test_create_agent():
-    """Test creating an Agent with application name."""
-    agent = Agent(FLM_TEST_APP, ctx=TestContext())
+def test_create_service_session():
+    """Test creating a service Session with application name."""
+    session = Session(FLM_TEST_APP, ctx=TestContext())
 
     ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.state == SessionState.OPEN]
     assert len(ssn_list) == 1
-    agent_id = agent.id()
-    assert ssn_list[0].id == agent_id
+    session_id = session.id()
+    assert ssn_list[0].id == session_id
     assert ssn_list[0].application == FLM_TEST_APP
     assert ssn_list[0].state == SessionState.OPEN
 
-    agent.close()
+    session.close()
 
     ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP]
     assert len(ssn_list) == 1
-    assert ssn_list[0].id == agent_id
+    assert ssn_list[0].id == session_id
     assert ssn_list[0].application == FLM_TEST_APP
     assert ssn_list[0].state == SessionState.CLOSED
 
 
 def test_invoke_task_without_context():
-    """Test invoking a task without common data using Agent API."""
-    agent = Agent(name=FLM_TEST_APP, ctx=None)
+    """Test invoking a task without common data using the service Session API."""
+    session = Session(name=FLM_TEST_APP, ctx=None)
 
     ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.state == SessionState.OPEN]
     assert len(ssn_list) == 1
-    assert ssn_list[0].id == agent.id()
+    assert ssn_list[0].id == session.id()
     assert ssn_list[0].application == FLM_TEST_APP
     assert ssn_list[0].state == SessionState.OPEN
 
     input = random_string()
 
-    output = agent.invoke(TestRequest(input=input))
+    output = session.invoke(TestRequest(input=input))
     assert output.output == input
     assert output.common_data is None
 
-    agent.close()
+    session.close()
 
 
 def test_invoke_task_with_context():
-    """Test invoking a task with context using Agent API."""
+    """Test invoking a task with context using the service Session API."""
     sys_context = random_string()
     input = random_string()
 
-    agent = Agent(FLM_TEST_APP, ctx=TestContext(common_data=sys_context))
+    session = Session(FLM_TEST_APP, ctx=TestContext(common_data=sys_context))
 
     ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.state == SessionState.OPEN]
     assert len(ssn_list) == 1
-    assert ssn_list[0].id == agent.id()
+    assert ssn_list[0].id == session.id()
     assert ssn_list[0].application == FLM_TEST_APP
     assert ssn_list[0].state == SessionState.OPEN
 
-    output = agent.invoke(TestRequest(input=input))
+    output = session.invoke(TestRequest(input=input))
     assert output.output == input
     assert output.common_data == sys_context
 
-    agent.close()
+    session.close()
 
 
 def test_update_context():
-    """Test updating context using Agent API."""
+    """Test updating context using the service Session API."""
     sys_context = random_string()
 
-    agent = Agent(FLM_TEST_APP, ctx=TestContext(common_data=sys_context))
+    session = Session(FLM_TEST_APP, ctx=TestContext(common_data=sys_context))
 
     ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.state == SessionState.OPEN]
     assert len(ssn_list) == 1
-    assert ssn_list[0].id == agent.id()
+    assert ssn_list[0].id == session.id()
     assert ssn_list[0].application == FLM_TEST_APP
     assert ssn_list[0].state == SessionState.OPEN
 
     previous_context = sys_context
     for _ in range(5):
         new_input_data = random_string()
-        output = agent.invoke(TestRequest(input=new_input_data, update_common_data=True))
+        output = session.invoke(TestRequest(input=new_input_data, update_common_data=True))
         assert output.output == new_input_data
         assert output.common_data == previous_context
 
-        ctx = agent.context()
+        ctx = session.context()
         assert ctx.common_data == new_input_data
 
         previous_context = new_input_data
 
-    agent.close()
+    session.close()
 
 
-def test_agent_context():
-    """Test getting Agent context."""
+def test_service_session_context():
+    """Test getting service Session context."""
     sys_context = random_string()
 
-    agent = Agent(FLM_TEST_APP, ctx=TestContext(common_data=sys_context))
+    session = Session(FLM_TEST_APP, ctx=TestContext(common_data=sys_context))
 
-    ctx = agent.context()
+    ctx = session.context()
     assert ctx is not None
     assert isinstance(ctx, TestContext)
     assert ctx.common_data == sys_context
 
-    agent.close()
+    session.close()
 
 
-def test_agent_context_manager():
-    """Test using Agent as a context manager."""
+def test_service_session_context_manager():
+    """Test using service Session as a context manager."""
     input = random_string()
 
-    with Agent(FLM_TEST_APP, ctx=None) as agent:
+    with Session(FLM_TEST_APP, ctx=None) as session:
         ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.state == SessionState.OPEN]
         assert len(ssn_list) == 1
-        agent_id = agent.id()
-        assert ssn_list[0].id == agent_id
+        session_id = session.id()
+        assert ssn_list[0].id == session_id
 
-        output = agent.invoke(TestRequest(input=input))
+        output = session.invoke(TestRequest(input=input))
         assert output.output == input
 
     # After context exit, session should be closed
-    ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.id == agent_id]
+    ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.id == session_id]
     assert len(ssn_list) == 1
-    assert ssn_list[0].id == agent_id
+    assert ssn_list[0].id == session_id
     assert ssn_list[0].state == SessionState.CLOSED
 
 
-def test_agent_open_existing_session():
-    """Test opening an existing session with Agent."""
+def test_service_session_open_existing_session():
+    """Test opening an existing session with service Session."""
     # First create a session and keep it open
-    agent1 = Agent(FLM_TEST_APP, ctx=TestContext(common_data=random_string()))
-    agent_id = agent1.id()
+    session1 = Session(FLM_TEST_APP, ctx=TestContext(common_data=random_string()))
+    session_id = session1.id()
 
-    # Now open the same open session with Agent
-    agent2 = Agent(session_id=agent_id)
+    # Now open the same open session with service Session
+    session2 = Session(session_id=session_id)
 
-    assert agent2.id() == agent_id
-    assert agent2.id() == agent1.id()
+    assert session2.id() == session_id
+    assert session2.id() == session1.id()
 
-    ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.id == agent_id]
+    ssn_list = [s for s in flamepy.list_sessions() if s.application == FLM_TEST_APP and s.id == session_id]
     assert len(ssn_list) == 1
-    assert ssn_list[0].id == agent_id
+    assert ssn_list[0].id == session_id
     assert ssn_list[0].state == SessionState.OPEN
 
-    agent1.close()
-    agent2.close()
+    session1.close()
+    session2.close()
 
 
-def test_agent_validation():
-    """Test that Agent raises ValueError when both name and session_id are None."""
+def test_service_session_validation():
+    """Test that service Session raises ValueError when both name and session_id are None."""
     with pytest.raises(ValueError, match="Either 'name' or 'session_id' must be provided"):
-        Agent(name=None, session_id=None)
+        Session(name=None, session_id=None)
 
 
-def test_agent_validation_both_provided():
-    """Test that Agent raises ValueError when both name and session_id are provided."""
+def test_service_session_validation_both_provided():
+    """Test that service Session raises ValueError when both name and session_id are provided."""
     with pytest.raises(ValueError, match="Cannot provide both 'name' and 'session_id'"):
-        Agent(name=FLM_TEST_APP, session_id="some-session-id")
+        Session(name=FLM_TEST_APP, session_id="some-session-id")
