+++ examples/agents/langchain/single_ssn.py
@@ -1,5 +1,4 @@
-import flamepy
-from flamepy.agent import Agent
+from flamepy.service import Session
 from apis import SysPrompt, Question
 
 LANGCHAIN_AGENT_NAME = "langchain-agent"
@@ -9,11 +8,11 @@ def ask_agent():
     sys_prompt = SysPrompt(prompt="You are a weather forecaster.")
     question = Question(question="Who are you?")
 
-    agent = Agent(LANGCHAIN_AGENT_NAME, ctx=sys_prompt)
-    output = agent.invoke(question)
+    session = Session(LANGCHAIN_AGENT_NAME, ctx=sys_prompt)
+    output = session.invoke(question)
 
     print(output.answer)
-    agent.close()
+    session.close()
 
 
 if __name__ == "__main__":
