+++ examples/agents/openai/client.py
@@ -12,8 +12,7 @@
 import argparse
 from typing import Optional
 
-import flamepy
-from flamepy.agent import Agent
+from flamepy.service import Session
 from apis import MyContext, Question
 
 OPENAI_APP_NAME = "openai-agent"
@@ -21,26 +20,26 @@ OPENAI_APP_NAME = "openai-agent"
 
 def main(message: str, ssn_id: Optional[str] = None):
     if ssn_id:
-        agent = Agent(session_id=ssn_id)
+        session = Session(session_id=ssn_id)
     else:
         sys_prompt = """You are a weather forecaster.
         If you are asked to fetch the weather, you should use the fetch_weather tool after confirming the location with the user.
         """
-        agent = Agent(OPENAI_APP_NAME, ctx=MyContext(prompt=sys_prompt))
+        session = Session(OPENAI_APP_NAME, ctx=MyContext(prompt=sys_prompt))
 
     print(f"{'=' * 30}")
-    print(f"Conversation <{agent.id()}>")
+    print(f"Conversation <{session.id()}>")
     print(f"{'=' * 30}")
 
     print(f"User: {message}")
 
-    output = agent.invoke(Question(question=message))
+    output = session.invoke(Question(question=message))
 
     print(f"Agent: {output.answer}")
 
-    cxt = agent.context()
+    cxt = session.context()
     print(f"{'=' * 30}")
-    print(f"Session History")
+    print("Session History")
     print(f"{'=' * 30}")
     if getattr(cxt, "messages", None) is not None:
         for msg in cxt.messages:
@@ -48,7 +47,7 @@ def main(message: str, ssn_id: Optional[str] = None):
     else:
         print("No history!")
 
-    agent.close()
+    session.close()
 
 
 if __name__ == "__main__":
