+++ examples/agents/sra/sra.py
@@ -1,5 +1,5 @@
 import flamepy
-from flamepy import agent
+from flamepy import service
 import os
 import qdrant_client
 from qdrant_client.models import VectorParams, Distance
@@ -156,7 +156,7 @@ async def collect_data(topic: str) -> list[str]:
     return payloads
 
 
-ins = agent.FlameInstance()
+ins = service.FlameInstance()
 
 sys_prompt = """
 You are a writer agent for research; you will write the research paper based on the research topics and the necessary information from the tools.
