+++ examples/crawler/crawler.py
@@ -1,42 +1,42 @@
+import io
+import re
+from urllib.parse import urljoin, urlparse
 
 import markitdown
-import flamepy
-from flamepy import agent
-import io
 import requests
-import re
 from bs4 import BeautifulSoup
-from urllib.parse import urlparse, urljoin
+from flamepy import service
 
-from apis import WebPage, Summary
+from apis import Summary, WebPage
 
-ins = agent.FlameInstance()
+ins = service.FlameInstance()
 
-headers = {
-    'User-Agent': 'Xflops Crawler 1.0',
-    'From': 'support@xflops.io'
-}
+headers = {"User-Agent": "Xflops Crawler 1.0", "From": "support@xflops.io"}
+
 
 def get_encoding(soup):
     encoding = None
     if soup:
         for meta_tag in soup.find_all("meta"):
-            encoding = meta_tag.get('charset')
-            if encoding: break
-            else:
-                encoding = meta_tag.get('content-type')
-                if encoding: break
-                else:
-                    content = meta_tag.get('content')
-                    if content:
-                        match = re.search('charset=(.*)', content)
-                        if match:
-                           encoding = match.group(1)
-                           break
+            encoding = meta_tag.get("charset")
+            if encoding:
+                break
+
+            encoding = meta_tag.get("content-type")
+            if encoding:
+                break
+
+            content = meta_tag.get("content")
+            if content:
+                match = re.search("charset=(.*)", content)
+                if match:
+                    encoding = match.group(1)
+                    break
     if encoding:
         return str(encoding).lower()
-    
+
     return "utf-8"
+
 
 @ins.entrypoint
 def crawler_app(wp: WebPage) -> Summary:
@@ -44,18 +44,18 @@ def crawler_app(wp: WebPage) -> Summary:
 
     soup = BeautifulSoup(text, "html.parser")
     encoding = get_encoding(soup)
-    
+
     links = []
     for link in soup.find_all("a"):
-        l = link.get("href")
-        if l is None:
+        href = link.get("href")
+        if href is None:
             continue
-        u = urlparse(l)
+        u = urlparse(href)
         if u.scheme == "http" or u.scheme == "https":
             links.append(u.geturl())
         elif u.scheme == "":
-            links.append(urljoin(wp.url, l))
-    
+            links.append(urljoin(wp.url, href))
+
     links = list(set(links))
 
     md = markitdown.MarkItDown()
@@ -63,6 +63,7 @@ def crawler_app(wp: WebPage) -> Summary:
     result = md.convert(stream).text_content
 
     return Summary(links=links, content=result)
+
 
 if __name__ == "__main__":
     ins.run()
