+++ sdk/python/src/flamepy/service/__init__.py
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from .client import Agent
+from .client import Session
 from .instance import FlameInstance
 
-__all__ = ["FlameInstance", "Agent"]
+__all__ = ["FlameInstance", "Session"]
